+++ CHANGELOG.md
@@ -16,6 +16,7 @@ develop
     - Revert “Use selection for Replace” action to allow using an empty string.
     - Tweak layout and style.
 - `cot` command now opens symbolic link target rather than the link itself.
+- Improve file encoding detection.
 
 
 ### Fixes